+++ NEWS.md
@@ -2,6 +2,12 @@
 
 * `bright_dark_period()` now maintains the date when looping the data.
 
+* Added articles on `Import & Cleaning`, `Metrics`, and `Visualizations` to the website.
+
+* Added the option for more print rows of observation intervals during `import`.
+
+* Added the option to set a length for the dataset starting from the end in `filter_Datetime()` and family.
+
 # LightLogR 0.3.5
 
 * Added the function `aggregate_Date()` to aggregate long datasets to one day per group.